requirements.txt
Django
psycopg2-binary
pillow
escpos
python-bidi
arabic-reshaper